Allow measurement window to drift + reset when keyed
Measurement window can now drift slowly up and down, to account for
environmental changes. A state machine that was present before but not
explicit replaces the spaghetti ifs that came before. At key, the
measurement window is reset to ensure that we know the current average
measurement and are able to respond to change.

We know that a person's touch would increase the capacitance, and thus
that when we are keyed, we expect our measurement to be less (not more)
than the average. This makes us sensitive to keys "press", but not to
"release".